ttys: Extract ttys_term.c (termious_ops)
